+++ CHANGELOG.md
@@ -4,6 +4,21 @@ All notable changes to [ChaiLove](https://github.com/RobLoach/ChaiLove) will be 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## 0.24.0 - 2018-09-03
+### Features
+- Update to use libretro-common audio
+- Added `love.filesystem.newFileData()`
+- Added `love.system.getClipboardText()`
+- Added `love.system.setClipboardText()`
+- Added `love.data.encode()`
+- Added `love.data.decode()`
+
+### Chores
+- Added `love.filesystem.mount()` test
+- Updated documentation
+- Updated unit testing framework
+- Additional method chaining functions
+
 ## 0.23.1 - 2018-08-28
 ### Fixes
 - Clean up of git submodules